Delete apiCalls.js, move apiCalls to userClass
