+++ src/index.js
@@ -1,3 +1,2 @@
 import './style.css';
 import './modules/form.js';
-import './modules/apiCall.js';
