+++ src/modules/useClass.js
@@ -1,7 +1,7 @@
 export default class User {
   constructor(name, score) {
-    this.name = name,
-    this.score = score
+    this.name = name;
+    this.score = score;
   }
 
   createUserScore = async () => {
@@ -19,4 +19,41 @@ export default class User {
       return error;
     }
   }
+
+  static createGame = async () => {
+  try {
+    const game = {
+      name: 'The Last Airbender'
+    }
+    const response = await fetch(`https://us-central1-js-capstone-backend.cloudfunctions.net/api/games`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body:JSON.stringify(game)
+  });
+  const data = await response.json();
+  return data;
+  } catch (error) {
+    return error;
+  }
+}
+
+  static fetchScores = async () => {
+    try {
+      const response = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/Pr8hkBfg3g5nFTZFaxmC/scores', 
+      {
+        method: 'GET',
+        headers: {
+          'Content-type': 'application/json'
+        }
+      });
+  
+      const data = await response.json();
+      console.log(data);
+      return data;
+    } catch (error) {
+      return error;
+    }
+  } 
 }
